Move MarkdownTranslator to new file
